Move into platform specific files, use GTK helper object, remove extra error message
