+++ interface/wx/filefn.h
@@ -363,6 +363,8 @@ bool wxRemoveFile(const wxString& file);
     @note On Unix, moving files across filesystem boundaries (e.g., from an
     external drive) is not supported and will return @false.
 
+    @availability{wxHAS_MOVE_TO_TRASH}
+
     @since 3.3.3
 
     @header{wx/filefn.h}
